fixing issue where database deletes old db values when rerunning

+++ database.js
@@ -7,28 +7,26 @@ const Database = require('better-sqlite3');
 const db = new Database('user.db');
 
 // Is the database initialized or do we need to initialize it?
-const stmt = db.prepare(`SELECT name FROM sqlite_master WHERE type='table' and name='userinfo';`);
+const stmt = db.prepare(`SELECT name FROM sqlite_master WHERE type='table' and name='usertable';`);
 let row = stmt.get();
-//if (row === undefined) {
+if (row === undefined) {
 // Echo information about what you are doing to the console.
     console.log('Your database appears to be empty. I will initialize it now.');
 // Set a const that will contain your SQL commands to initialize the database.
     const sqlInit = `
-        DROP TABLE userinfo;
-        CREATE TABLE userinfo ( group_id INTEGER PRIMARY KEY AUTOINCREMENT, user TEXT, pass TEXT, email TEXT UNIQUE, CONSTRAINT email_unique UNIQUE (email) );
-		INSERT INTO userinfo (user, pass, email) VALUES ('admin','1234','123email@gmail.com'), ('test','5678','456email@gmail.com');
+        CREATE TABLE IF NOT EXISTS usertable ( group_id INTEGER PRIMARY KEY AUTOINCREMENT, user TEXT, pass TEXT, email TEXT UNIQUE, CONSTRAINT email_unique UNIQUE (email) );
+		INSERT INTO usertable (user, pass, email) VALUES ('admin','1234','123email@gmail.com'), ('test','5678','456email@gmail.com');
         DROP TABLE interactions;
-        CREATE TABLE interactions ( id INTEGER PRIMARY KEY AUTOINCREMENT, login INTEGER, inventory INTEGER, coins INTEGER, grp_id INTEGER NOT NULL, FOREIGN KEY (grp_id) REFERENCES userinfo(group_id) ON UPDATE CASCADE ON DELETE CASCADE );
+        CREATE TABLE interactions ( id INTEGER PRIMARY KEY AUTOINCREMENT, login INTEGER, inventory INTEGER, coins INTEGER, grp_id INTEGER NOT NULL, FOREIGN KEY (grp_id) REFERENCES usertable (group_id) ON UPDATE CASCADE ON DELETE CASCADE );
         INSERT INTO interactions (login, inventory, coins, grp_id) VALUES (0, 0, 0, 1);
         `;
-    //`;
 // Execute SQL commands that we just wrote above.
     db.exec(sqlInit);
 // Echo information about what we just did to the console.
     console.log('Database has been initialized with two tables, one for users and one for interactions.');
-//} else {
+} else {
 // Since the database already exists, echo that to the console.
-//    console.log('Database exists!')
-//}
+    console.log('Database exists!')
+}
 // Export all of the above as a module so that we can use it elsewhere.
 module.exports = db
